requirements.txt
# Core Dependencies
aiogram==3.4.1
python-dotenv==1.0.0

# HTTP & API
httpx==0.26.0

# Security
cryptography==41.0.7

# Data Processing
python-dateutil==2.8.2

# ==================== PostgreSQL ====================
# Database ORM and driver
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Database migrations
alembic==1.13.1

# ==================== Development & Testing ====================
# Testing framework
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
